\name{CALINE4-package}
\alias{CALINE4-package}
\alias{CALINE4}
\docType{package}
\title{CALINE4}
\description{CALINE4}
\details{
\tabular{ll}{
Package: \tab CALINE4\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-06-10\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr
}
}
\author{David Holstius}
\keyword{package}
\keyword{CALINE}
\keyword{CALINE4}
